fix 9342 API endpoint changes
fix #9342 API endpoint changes ; Split Gene Panel into Meta and Data, Added Generic assay Data. Added tests for all.
