+++ web/src/main/java/org/cbioportal/web/config/PublicApiTags.java
@@ -15,6 +15,7 @@ public class PublicApiTags {
     public static final String COPY_NUMBER_SEGMENTS = "Copy Number Segments";
     public static final String GENES = "Genes";
     public static final String GENE_PANELS = "Gene Panels";
+    public static final String GENE_PANEL_DATA = "Gene Panel Data";
     public static final String GENERIC_ASSAYS = "Generic Assays";
     public static final String GENERIC_ASSAY_DATA = "Generic Assay Data";
     public static final String INFO = "Info";
