+++ web/src/test/java/org/cbioportal/web/GenePanelControllerTest.java
@@ -164,8 +164,7 @@ public class GenePanelControllerTest {
             .andExpect(MockMvcResultMatchers.status().isOk())
             .andExpect(MockMvcResultMatchers.header().string(HeaderKeyConstants.TOTAL_COUNT, "2"));
     }
-
-
+    
     @Test
     public void getGenePanelNotFound() throws Exception {
 
@@ -214,98 +213,5 @@ public class GenePanelControllerTest {
             .andExpect(MockMvcResultMatchers.jsonPath("$.genes[1].entrezGeneId").value(TEST_ENTREZ_GENE_ID_2))
             .andExpect(MockMvcResultMatchers.jsonPath("$.genes[1].hugoGeneSymbol").value(TEST_HUGO_GENE_SYMBOL_2));
     }
-
-    @Test
-    public void getGenePanelData() throws Exception {
-
-        List<GenePanelData> genePanelDataList = createExampleGenePanelData();
-
-        Mockito.when(genePanelService.getGenePanelData(Mockito.anyString(), Mockito.anyString())).thenReturn(genePanelDataList);
-
-        GenePanelDataFilter genePanelDataFilter = new GenePanelDataFilter();
-        genePanelDataFilter.setSampleListId(TEST_SAMPLE_LIST_ID);
-
-        mockMvc.perform(MockMvcRequestBuilders.post(
-            "/molecular-profiles/test_molecular_profile_id/gene-panel-data/fetch")
-            .accept(MediaType.APPLICATION_JSON)
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(objectMapper.writeValueAsString(genePanelDataFilter)))
-            .andExpect(MockMvcResultMatchers.status().isOk())
-            .andExpect(MockMvcResultMatchers.content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-            .andExpect(MockMvcResultMatchers.jsonPath("$", Matchers.hasSize(2)))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].molecularProfileId").value(TEST_MOLECULAR_PROFILE_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].sampleId").value(TEST_SAMPLE_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].genePanelId").value(TEST_GENE_PANEL_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].patientId").value(TEST_PATIENT_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].studyId").value(TEST_STUDY_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].profiled").value(true))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].molecularProfileId").value(TEST_MOLECULAR_PROFILE_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].sampleId").value(TEST_SAMPLE_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].genePanelId").doesNotExist())
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].patientId").value(TEST_PATIENT_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].studyId").value(TEST_STUDY_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].profiled").value(true));
-    }
-
-    @Test
-    public void fetchGenePanelData() throws Exception {
-
-        List<GenePanelData> genePanelDataList = createExampleGenePanelData();
-
-        Mockito.when(genePanelService.fetchGenePanelDataInMultipleMolecularProfiles(Mockito.anyList())).thenReturn(genePanelDataList);
-
-        List<SampleMolecularIdentifier> sampleMolecularIdentifiers = new ArrayList<>();
-        SampleMolecularIdentifier sampleMolecularIdentifier1 = new SampleMolecularIdentifier();
-        sampleMolecularIdentifier1.setMolecularProfileId(TEST_MOLECULAR_PROFILE_ID_1);
-        sampleMolecularIdentifier1.setSampleId(TEST_SAMPLE_ID_1);
-        sampleMolecularIdentifiers.add(sampleMolecularIdentifier1);
-        SampleMolecularIdentifier sampleMolecularIdentifier2 = new SampleMolecularIdentifier();
-        sampleMolecularIdentifier2.setMolecularProfileId(TEST_MOLECULAR_PROFILE_ID_2);
-        sampleMolecularIdentifier2.setSampleId(TEST_SAMPLE_ID_2);
-        sampleMolecularIdentifiers.add(sampleMolecularIdentifier2);
-        GenePanelDataMultipleStudyFilter genePanelDataMultipleStudyFilter = new GenePanelDataMultipleStudyFilter();
-        genePanelDataMultipleStudyFilter.setSampleMolecularIdentifiers(sampleMolecularIdentifiers);
-
-        mockMvc.perform(MockMvcRequestBuilders.post(
-            "/gene-panel-data/fetch")
-            .accept(MediaType.APPLICATION_JSON)
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(objectMapper.writeValueAsString(genePanelDataMultipleStudyFilter)))
-            .andExpect(MockMvcResultMatchers.status().isOk())
-            .andExpect(MockMvcResultMatchers.content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-            .andExpect(MockMvcResultMatchers.jsonPath("$", Matchers.hasSize(2)))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].molecularProfileId").value(TEST_MOLECULAR_PROFILE_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].sampleId").value(TEST_SAMPLE_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].genePanelId").value(TEST_GENE_PANEL_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].patientId").value(TEST_PATIENT_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].studyId").value(TEST_STUDY_ID_1))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[0].profiled").value(true))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].molecularProfileId").value(TEST_MOLECULAR_PROFILE_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].sampleId").value(TEST_SAMPLE_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].genePanelId").doesNotExist())
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].patientId").value(TEST_PATIENT_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].studyId").value(TEST_STUDY_ID_2))
-            .andExpect(MockMvcResultMatchers.jsonPath("$[1].profiled").value(true));
-    }
-
-    private List<GenePanelData> createExampleGenePanelData() {
-
-        List<GenePanelData> genePanelDataList = new ArrayList<>();
-        GenePanelData genePanelData1 = new GenePanelData();
-        genePanelData1.setGenePanelId(TEST_GENE_PANEL_ID_1);
-        genePanelData1.setSampleId(TEST_SAMPLE_ID_1);
-        genePanelData1.setMolecularProfileId(TEST_MOLECULAR_PROFILE_ID_1);
-        genePanelData1.setPatientId(TEST_PATIENT_ID_1);
-        genePanelData1.setStudyId(TEST_STUDY_ID_1);
-        genePanelData1.setProfiled(true);
-        genePanelDataList.add(genePanelData1);
-        GenePanelData genePanelData2 = new GenePanelData();
-        genePanelData2.setSampleId(TEST_SAMPLE_ID_2);
-        genePanelData2.setMolecularProfileId(TEST_MOLECULAR_PROFILE_ID_2);
-        genePanelData2.setPatientId(TEST_PATIENT_ID_2);
-        genePanelData2.setStudyId(TEST_STUDY_ID_2);
-        genePanelData2.setProfiled(true);
-        genePanelDataList.add(genePanelData2);
-        return genePanelDataList;
-    }
+    
 }
